requirements.txt
# Core dependencies
anthropic>=0.40.0
duckdb>=1.1.0
pandas>=2.2.0
pyarrow>=17.0.0
python-dotenv>=1.0.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0

# UI
streamlit>=1.31.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
